Activity 3/requirements.txt

psycopg2-binary==2.9.1
